0.0.8 Add offline picture feature.
Offline picture:
When the UIWebView receive an url to load the class try to find it in
the temporary directory. If there isn’t, the class download it from the
web and store it.
When the tableview update his content with new data, if this data comes
from the web (isOnline == true) than the temporary directory is
cleaned, ready to receive the new pictures.
So if the app is used offline the tableview can show the pictures
downloaded in the last online session.

DetailView :
Added an “open in safari” button on the bottom toolbar.

Performed some UI general adjustment for a better UX